# .static/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a server subscription API.",
        "title": "Subscription API",
        "contact": {
            "name": "Maksim Braer",
            "email": "braer.maks@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/stats/total": {
            "get": {
                "description": "Calculate total cost of subscriptions for the period. Date format: MM-YYYY (e.g., \"01-2024\", \"12-2024\")",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stats"
                ],
                "summary": "Get total stats",
                "parameters": [
                    {
                        "type": "string",
                        "example": "550e8400-e29b-41d4-a716-446655440000",
                        "description": "User UUID",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "Netflix",
                        "description": "Service name",
                        "name": "service_name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "01-2024",
                        "description": "Period start (MM-YYYY)",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "12-2024",
                        "description": "Period end (MM-YYYY)",
                        "name": "end_date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.GetTotalStatsResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid arguments or date format",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "invalid arguments"
                }
            }
        },
        "handlers.Filters": {
            "type": "object",
            "properties": {
                "service_name": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "handlers.GetTotalStatsResponse": {
            "type": "object",
            "properties": {
                "filters": {
                    "$ref": "#/definitions/handlers.Filters"
                },
                "period": {
                    "$ref": "#/definitions/handlers.Period"
                },
                "subscriptions_count": {
                    "type": "integer"
                },
                "total_cost": {
                    "type": "integer"
                }
            }
        },
        "handlers.Period": {
            "type": "object",
            "properties": {
                "end": {
                    "type": "string"
                },
                "start": {
                    "type": "string"
                }
            }
        },
...
    }
}